text2speech/requirements.txt
torch==1.6.0
six
typeguard
soundfile
pyyaml
h5py
scipy